backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb
groq==0.4.2
pypdf==3.17.4
Pillow==10.1.0
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
